+++ packages/nimble-components/src/table/components/cell/tests/table-cell.spec.ts
@@ -28,7 +28,7 @@ class TestTableColumnCellView extends TableCellView<SimpleTableCellRecord> {}
 
 // prettier-ignore
 async function setup(): Promise<Fixture<TableCell<SimpleTableCellRecord>>> {
-    return fixture<TableCell<SimpleTableCellRecord>>(
+    return await fixture<TableCell<SimpleTableCellRecord>>(
         html`<nimble-table-cell>
             </nimble-table-cell>`
     );