+++ packages/nimble-components/src/wafer-map/tests/hover-handler.spec.ts
@@ -7,7 +7,7 @@ import { processUpdates } from '../../testing/async-helpers';
 import { Fixture, fixture } from '../../utilities/tests/fixture';
 
 async function setup(): Promise<Fixture<WaferMap>> {
-    return fixture<WaferMap>(html`<nimble-wafer-map></nimble-wafer-map>`);
+    return await fixture<WaferMap>(html`<nimble-wafer-map></nimble-wafer-map>`);
 }
 
 describe('HoverHandler', () => {